+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   used when reading an associated `Nucleus::Bytes`. When `Nucleus::Bytes`
   contains `BufferedBytes::Scratch`, `scratch` will contain the bytes contained
   in the atom.
+- `SymbolMapRef` is now a struct with private contents.
 
 ### Changed